replace all doubles with floats

+++ src/activation.cpp
@@ -4,39 +4,39 @@
 
 #include "matrix.h"
 
-double relu_single(double x) {
+float relu_single(float x) {
     return x > 0 ? x : 0;
 }
 
-void relu(std::vector<double>& x) {
+void relu(std::vector<float>& x) {
     vec_apply(x, relu_single);
 }
 
-double relu_derivative_single(double x) {
+float relu_derivative_single(float x) {
     return x > 0 ? 1 : 0;
 }
 
-void relu_derivative(std::vector<double>& x) {
+void relu_derivative(std::vector<float>& x) {
     vec_apply(x, relu_derivative_single);
 }
 
-void softmax(std::vector<double>& x) {
-    double sum = 0;
-    for (double i : x) {
+void softmax(std::vector<float>& x) {
+    float sum = 0;
+    for (float i : x) {
         sum += exp(i);
     }
-    for (double & i : x) {
+    for (float & i : x) {
         i = exp(i) / sum;
     }
 }
 
-void softmax_derivative(std::vector<double> &x) {
+void softmax_derivative(std::vector<float> &x) {
     for (size_t i = 0; i < x.size(); i++) {
         x[i] = 1;
     }
 }
 
-std::function<void(std::vector<double>&)> get_activation(Activation activation) {
+std::function<void(std::vector<float>&)> get_activation(Activation activation) {
     switch (activation) {
         case Activation::RELU:
             return relu;
@@ -46,7 +46,7 @@ std::function<void(std::vector<double>&)> get_activation(Activation activation) 
     assert(false);
 }
 
-std::function<void(std::vector<double>&)> get_activation_derivative(Activation activation) {
+std::function<void(std::vector<float>&)> get_activation_derivative(Activation activation) {
     switch (activation) {
         case Activation::RELU:
             return relu_derivative;
