+++ src/matrix.cpp
@@ -5,7 +5,7 @@
 
 #include "matrix.h"
 
-const double& matrix_row_view::operator[](size_t index) const {
+const float& matrix_row_view::operator[](size_t index) const {
     return data[index];
 }
 
@@ -15,7 +15,7 @@ matrix::matrix(int rows, int cols) {
     data.resize(rows * cols, 0);
 }
 
-matrix::matrix(std::vector<double> data, int rows, int cols) {
+matrix::matrix(std::vector<float> data, int rows, int cols) {
     this->rows = rows;
     this->cols = cols;
     this->data = std::move(data);
@@ -47,12 +47,12 @@ void matrix::zero() {
     std::fill(data.begin(), data.end(), 0);
 }
 
-std::vector<double> matrix::calc_potentials(const matrix_row_view &vec) {
-    std::vector<double> result;
+std::vector<float> matrix::calc_potentials(const matrix_row_view &vec) {
+    std::vector<float> result;
     result.reserve(rows);
 
     for (int i = 0; i < rows; i++) {
-        double sum = 0;
+        float sum = 0;
         for (int j = 0; j < cols - 1; j++) {
             sum += data[index(i, j)] * vec[j];
         }
@@ -65,15 +65,15 @@ std::vector<double> matrix::calc_potentials(const matrix_row_view &vec) {
     return result;
 }
 
-void vec_apply(std::vector<double> &vec, const std::function<double(double)>& func) {
+void vec_apply(std::vector<float> &vec, const std::function<float(float)>& func) {
     for (size_t i = 0; i < vec.size(); i++) {
         vec[i] = func(vec[i]);
     }
 }
 
-std::vector<double> vec_elementwise_mul(std::vector<double>& vec1, std::vector<double>& vec2) {
+std::vector<float> vec_elementwise_mul(std::vector<float>& vec1, std::vector<float>& vec2) {
     assert(vec1.size() == vec2.size());
-    std::vector<double> result;
+    std::vector<float> result;
     result.reserve(vec1.size());
     for (size_t i = 0; i < vec1.size(); i++) {
         result.push_back(vec1[i] * vec2[i]);
@@ -81,7 +81,7 @@ std::vector<double> vec_elementwise_mul(std::vector<double>& vec1, std::vector<d
     return result;
 }
 
-double sample_normal_dist(double mean, double stddev) {
+float sample_normal_dist(float mean, float stddev) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::normal_distribution<> dist(mean, stddev);
@@ -101,11 +101,11 @@ matrix outer_product(const matrix_row_view &vec1, const matrix_row_view &vec2) {
     return result;
 }
 
-std::vector<double> compute_de_dy(const std::vector<double> &prev_de_dy, const std::vector<double> &potential_der, const matrix &weights) {
-    std::vector<double> res(weights.cols - 1, 0);
+std::vector<float> compute_de_dy(const std::vector<float> &prev_de_dy, const std::vector<float> &potential_der, const matrix &weights) {
+    std::vector<float> res(weights.cols - 1, 0);
 
     for (size_t i = 0; i < prev_de_dy.size(); i++) {
-        double factor = prev_de_dy[i] * potential_der[i];
+        float factor = prev_de_dy[i] * potential_der[i];
         // skip bias
         for (int j = 0; j < weights.cols - 1; j++) {
             res[j] += factor * weights.data[weights.index(static_cast<int>(i), j)];
@@ -123,14 +123,15 @@ inline int matrix::index(int row, int col) const {
 }
 
 void matrix::normalize_data() {
+    // TODO normalize by column
     // subtract mean and divide by standard deviation
     for (int i = 0; i < cols - 1; i++) {
-        double sum = 0;
+        float sum = 0;
         for (int j = 0; j < rows; j++) {
             sum += data[index(j, i)];
         }
-        double mean = sum / rows;
-        double stddev = 0;
+        float mean = sum / rows;
+        float stddev = 0;
         for (int j = 0; j < rows; j++) {
             stddev += (data[index(j, i)] - mean) * (data[index(j, i)] - mean);
         }
@@ -141,13 +142,13 @@ void matrix::normalize_data() {
     }
 }
 
-void matrix::multiply_by_scalar(double scalar) {
+void matrix::multiply_by_scalar(float scalar) {
     for (int i = 0; i < size(); i++) {
         data[i] *= scalar;
     }
 }
 
-void add_scaled_vector(std::vector<double> &vec, const std::vector<double> &vec_to_add, double scale) {
+void add_scaled_vector(std::vector<float> &vec, const std::vector<float> &vec_to_add, float scale) {
     for (size_t i = 0; i < vec.size(); i++) {
         vec[i] += vec_to_add[i] * scale;
     }
